Implemented a basic framework for use by new experiments that will be executed using Shady.
The framework offers easy local development: all the files are simple
HTML, CSS and JavaScript files that can be edited with the use of an
easy to setup webserver like lighttp and co.

Deployment is also easy, currently the developer does not need to take
any action. Shady will supply necessary environment variables.